Move four more Enterprise-only controllers into pkg/enterprise
Application layer, egress gateway, packet capture and OpenTelemetry collector,
with the application layer and egress gateway render packages.
