+++ internal/controller/controllers.go
@@ -73,13 +73,6 @@ func AddToManager(mgr ctrl.Manager, options options.ControllerOptions) error {
 	}).SetupWithManager(mgr, options); err != nil {
 		return fmt.Errorf("failed to create controller Istio: %v", err)
 	}
-	if err := (&ApplicationLayerReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("ApplicationLayer"),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr, options); err != nil {
-		return fmt.Errorf("failed to create controller %s: %v", "ApplicationLayer", err)
-	}
 	if err := (&ManagerReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("Manager"),
@@ -122,13 +115,6 @@ func AddToManager(mgr ctrl.Manager, options options.ControllerOptions) error {
 	}).SetupWithManager(mgr, options); err != nil {
 		return fmt.Errorf("failed to create controller %s: %v", "PolicyRecommendation", err)
 	}
-	if err := (&EgressGatewayReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("EgressGateway"),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr, options); err != nil {
-		return fmt.Errorf("failed to create controller %s: %v", "EgressGateway", err)
-	}
 	if err := (&SecretsReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("Secrets"),
@@ -149,13 +135,6 @@ func AddToManager(mgr ctrl.Manager, options options.ControllerOptions) error {
 		Scheme: mgr.GetScheme(),
 	}).SetupWithManager(mgr, options); err != nil {
 		return fmt.Errorf("failed to create controller %s: %v", "CSR", err)
-	}
-	if err := (&PacketCaptureReconciler{
-		Client: mgr.GetClient(),
-		Log:    ctrl.Log.WithName("controllers").WithName("PacketCapture"),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr, options); err != nil {
-		return fmt.Errorf("failed to create controller %s: %v", "PacketCapture", err)
 	}
 	if err := (&GatewayAPIReconciler{
 		Client: mgr.GetClient(),
@@ -187,12 +166,6 @@ func AddToManager(mgr ctrl.Manager, options options.ControllerOptions) error {
 		Scheme: mgr.GetScheme(),
 	}).SetupWithManager(mgr, options); err != nil {
 		return fmt.Errorf("failed to create controller %s: %v", "PodIPRecovery", err)
-	}
-	if err := (&OpenTelemetryCollectorReconciler{
-		Client: mgr.GetClient(),
-		Scheme: mgr.GetScheme(),
-	}).SetupWithManager(mgr, options); err != nil {
-		return fmt.Errorf("failed to create controller %s: %v", "OpenTelemetry", err)
 	}
 	// +kubebuilder:scaffold:builder
