+++ pkg/enterprise/controller/otelcollector/otelcollector_controller_test.go
