+++ pkg/enterprise/controller/otelcollector/otelcollector_suite_test.go
@@ -1,10 +1,10 @@
-// Copyright (c) 2021-2026 Tigera, Inc. All rights reserved.
+// Copyright (c) 2026 Tigera, Inc. All rights reserved.
 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,18 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package applicationlayer
+package otelcollector
 
 import (
 	"testing"
 
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
+	uzap "go.uber.org/zap"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
 
-func TestStatus(t *testing.T) {
+func TestController(t *testing.T) {
+	logf.SetLogger(zap.New(zap.WriteTo(ginkgo.GinkgoWriter), zap.UseDevMode(true), zap.Level(uzap.NewAtomicLevelAt(uzap.DebugLevel))))
 	gomega.RegisterFailHandler(ginkgo.Fail)
 	suiteConfig, reporterConfig := ginkgo.GinkgoConfiguration()
-	reporterConfig.JUnitReport = "../../../report/ut/applicationlayer_controller_suite.xml"
-	ginkgo.RunSpecs(t, "pkg/controller/applicationlayer Controller Suite", suiteConfig, reporterConfig)
+	reporterConfig.JUnitReport = "../../../report/ut/otelcollector_controller_suite.xml"
+	ginkgo.RunSpecs(t, "pkg/enterprise/controller/otelcollector Controller Suite", suiteConfig, reporterConfig)
 }
