+++ pkg/enterprise/controller/packetcapture/packetcapture_controller.go
@@ -56,11 +56,6 @@ var log = logf.Log.WithName("controller_packet_capture")
 // Add creates a new PacketCapture Controller and adds it to the Manager. The Manager will set fields on the Controller
 // and Start it when the Manager is Started.
 func Add(mgr manager.Manager, opts options.ControllerOptions) error {
-	if !opts.Variant.IsEnterprise() {
-		// No need to start this controller
-		return nil
-	}
-
 	tierWatchReady := &utils.ReadyFlag{}
 
 	r := newReconciler(mgr, opts, tierWatchReady)
