+++ pkg/enterprise/controllers_test.go
@@ -30,7 +30,7 @@ var _ = Describe("Controllers", func() {
 				Expect(c.Add).NotTo(BeNil())
 				names = append(names, c.Name)
 			}
-			Expect(names).To(ContainElement("Monitor"))
+			Expect(names).To(ConsistOf("Monitor", "ApplicationLayer", "EgressGateway", "PacketCapture", "OpenTelemetryCollector"))
 		},
 		Entry("CalicoEnterprise", operatorv1.CalicoEnterprise),
 		//nolint:staticcheck // SA1019: the deprecated spelling is what this covers
