+++ pkg/enterprise/installation/kubecontrollers.go
@@ -40,9 +40,9 @@ import (
 	"github.com/tigera/operator/pkg/controller/utils/imageset"
 	"github.com/tigera/operator/pkg/dns"
 	entkubecontrollers "github.com/tigera/operator/pkg/enterprise/kubecontrollers"
+	"github.com/tigera/operator/pkg/enterprise/render/applicationlayer"
 	"github.com/tigera/operator/pkg/extensions"
 	"github.com/tigera/operator/pkg/render"
-	"github.com/tigera/operator/pkg/render/applicationlayer"
 	rcomp "github.com/tigera/operator/pkg/render/common/components"
 	rmeta "github.com/tigera/operator/pkg/render/common/meta"
 	"github.com/tigera/operator/pkg/render/common/networkpolicy"
