+++ pkg/enterprise/render/applicationlayer/envoy-config.yaml.template
