+++ pkg/enterprise/render/applicationlayer/ruleset/coreruleset/tigera.conf
