+++ pkg/enterprise/render/egressgateway/egressgateway.go
